pyproject.toml
[tool.poetry]
name = "demo-django-celery-results"
version = "0.1.0"
description = ""
authors = ["Matthieu Etchegoyen <etchegom@gmail.com>"]

[tool.poetry.dependencies]
python = "3.7"
celery = "^4.3.0"
django = "2.2"
psycopg2-binary = "^2.8.4"
redis = "^3.3.11"
django-environ = "^0.4.5"
django-celery-results = "^1.1.2"
django-extensions = "^2.2.5"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
flake8 = "^3.7.9"
pytest = "^5.3.1"
rope = "^0.14.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
